pyproject.toml
[tool.poetry]
name = "chargepal-camera-kit"
version = "1.3.1"
description = "A tool kit for working with serial cameras."
authors = ["Georg John <georg.john@dfki.de>"]
readme = "README.md"
packages = [{include = "camera_kit"}]

[tool.poetry.dependencies]
python = "^3.8, <3.12"
opencv-python = "4.7.0.68"
pyrealsense2 = "^2.54.2.5684"
tomli = "^2.0.1"
tomli-w = "^1.0.0"
tomlkit = "^0.12.3"
tqdm = "^4.66.1"
pyyaml = "^6.0.1"
scipy = "1.9.3"
spatialmath-python = "^1.1.9"

[tool.poetry.group.dev.dependencies]
mypy = "^1.7.1"
types-tqdm = "^4.66.0.5"
types-pyyaml = "^6.0.12.12"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocs = "*"
mkdocs-material = "*"
mkdocstrings = "*"
mkdocstrings-python = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
